API/requirements.txt
fastapi
pydantic
supabase
python-dotenv
uvicorn